+++ app/assets/javascript/ahoy_captain/controllers/application_controller.js
@@ -10,6 +10,15 @@ export default class extends Controller {
         }, 1000 * 30);
       });
     }
+
+    document.querySelectorAll('a[data-turbo-frame]').forEach(link => {
+      const frameSelector = link.dataset.turboFrame;
+      const frame = document.querySelector(`turbo-frame#${frameSelector}`);
+      const src = frame.src;
+      if(link.href.includes(src)) {
+        link.classList.add('text-primary', 'font-semibold')
+      }
+    })
   }
 
   comboboxInit(event) {