+++ app/controllers/ahoy_captain/locations/maps_controller.rb
@@ -1,6 +1,8 @@
 module AhoyCaptain
   module Locations
     class MapsController < ApplicationController
+      include Limitable
+
       before_action do
         if Widget.disabled?(:locations, :map)
           raise Widget::WidgetDisabled.new("Widget disabled", :geography)
@@ -8,7 +10,14 @@ module AhoyCaptain
       end
 
       def show
-        @visits = visit_query.group("country").count
+        if request.variant.include?(:details)
+          results = CountryQuery.call(params)
+          results = results.limit(limit)
+          @countries = paginate(results).map { |country| CountryDecorator.new(country, self) }
+          render template: 'ahoy_captain/locations/countries/index'
+        else
+          @countries = visit_query.group("country").count
+        end
       end
     end
   end