+++ app/assets/javascript/ahoy_captain/controllers/predicate_select_controller.js
@@ -4,6 +4,7 @@ export default class extends Controller {
   static targets = ["select"]
 
   handleChange(event) {
+    this.selectTarget.dataset.predicate = event.target.value
     this.selectTarget.name = event.target.value
   }
 }