+++ app/components/ahoy_captain/previous_next_component.rb
@@ -1,0 +1,11 @@
+module AhoyCaptain
+  class PreviousNextComponent < ViewComponent::Base
+    def initialize(range)
+      @range = range
+    end
+
+    def render?
+      false
+    end
+  end
+end