+++ app/components/ahoy_captain/sticky_nav_component.rb
@@ -10,6 +10,14 @@ class AhoyCaptain::StickyNavComponent < ViewComponent::Base
     @filters ||= ::AhoyCaptain::FilterParser.parse(request)
   end
 
+  def custom_range_label
+    if range.custom?
+      [range.starts_at, range.ends_at].map { |date| date.strftime('%b %d, %Y') }.join("-")
+    else
+      "Custom Range"
+    end
+  end
+
   def tag_list_hidden?
     filters.values.map(&:values).flatten.size < ::AhoyCaptain::FilterParser::FILTER_MENU_MAX_SIZE
   end