+++ app/components/ahoy_captain/dropdown_link_component.rb
@@ -9,6 +9,10 @@ class AhoyCaptain::DropdownLinkComponent < ViewComponent::Base
     @classes = classes
   end
 
+  def link_to(name, url, **options)
+    self.with_option_content view_context.link_to name, url, options
+  end
+
   private
 
   attr_reader :title, :classes